backend/requirements-test.txt
pytest>=7.0
pytest-asyncio>=0.22
httpx>=0.24
pytest-django>=4.5
coverage
mypy
